perf(eth): zip confirmation variables
