+++ Cargo.toml
@@ -46,7 +46,6 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 structopt = "0.3.17"
 uint = "0.8.3"
-web3 = { git = "https://github.com/tomusdrw/rust-web3" }
 
 [dependencies.scale]
 package = "parity-scale-codec"
